moved assemblers, weavers to jbm-flow
