+++ jbpm-flow-builder/src/main/java/org/jbpm/assembler/BPMN2AssemblerService.java
@@ -14,19 +14,21 @@
  * limitations under the License.
  */
 
-package org.jbpm.process.assembler;
+package org.jbpm.assembler;
 
 import org.drools.compiler.builder.impl.KnowledgeBuilderImpl;
+import org.drools.compiler.compiler.BPMN2ProcessFactory;
 import org.kie.api.io.ResourceType;
 
-public class DRFAssemblerService extends AbstractProcessAssembler {
+public class BPMN2AssemblerService extends AbstractProcessAssembler {
 
     @Override
     public ResourceType getResourceType() {
-        return ResourceType.DRF;
+        return ResourceType.BPMN2;
     }
 
     @Override
     protected void configurePackageBuilder(KnowledgeBuilderImpl kb) {
+        BPMN2ProcessFactory.configurePackageBuilder(kb);
     }
 }
