+++ jbpm-flow-builder/src/main/java/org/jbpm/assembler/DRFAssemblerService.java
@@ -14,21 +14,19 @@
  * limitations under the License.
  */
 
-package org.jbpm.process.assembler;
+package org.jbpm.assembler;
 
 import org.drools.compiler.builder.impl.KnowledgeBuilderImpl;
-import org.drools.compiler.compiler.BPMN2ProcessFactory;
 import org.kie.api.io.ResourceType;
 
-public class BPMN2AssemblerService extends AbstractProcessAssembler {
+public class DRFAssemblerService extends AbstractProcessAssembler {
 
     @Override
     public ResourceType getResourceType() {
-        return ResourceType.BPMN2;
+        return ResourceType.DRF;
     }
 
     @Override
     protected void configurePackageBuilder(KnowledgeBuilderImpl kb) {
-        BPMN2ProcessFactory.configurePackageBuilder(kb);
     }
 }
