+++ jbpm-flow/src/main/resources/META-INF/kie.conf
@@ -1,6 +1,2 @@
 org.drools.core.marshalling.impl.ProcessMarshallerFactoryService=org.jbpm.marshalling.impl.ProcessMarshallerFactoryServiceImpl
 org.drools.core.runtime.process.ProcessRuntimeFactoryService=org.jbpm.process.instance.ProcessRuntimeFactoryServiceImpl
-org.kie.api.internal.assembler.KieAssemblers = +org.jbpm.process.assembler.BPMN2AssemblerService
-org.kie.api.internal.weaver.KieWeavers = +org.jbpm.process.weaver.BPMN2WeaverService
-org.kie.api.internal.assembler.KieAssemblers = +org.jbpm.process.assembler.DRFAssemblerService
-org.kie.api.internal.weaver.KieWeavers = +org.jbpm.process.weaver.DRFWeaverService
